% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site-seq.R
\name{add_site_seq}
\alias{add_site_seq}
\title{Add site-specific sequence information.}
\usage{
add_site_seq(exp, fasta, n_aa = 7)
}
\arguments{
\item{exp}{A \code{glyexp::experiment()} object.}

\item{fasta}{A character string specifying the path to the FASTA file containing protein sequences.}

\item{n_aa}{The number of amino acids to the left and right of the glycosylation site.
For example, if \code{n_aa = 5}, the resulting sequence will contain 11 amino acids.}
}
\value{
A \code{glyexp::experiment()} object with the new "site_sequence" column.
}
\description{
This function adds a new "site_sequence" column to the variable information table.
It is the protein sequence around the glycosylation site.
If the head and tail amino acids of the peptide sequence are insufficient, fill with "X".

This function requires the following columns in the variable information tibble:
\itemize{
\item "protein": The protein uniprot accession.
\item "protein_site": The site on the protein sequence.
}
}
