% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto-correct-batch-effect.R
\name{auto_correct_batch_effect}
\alias{auto_correct_batch_effect}
\title{Automatic Batch Correction}
\usage{
auto_correct_batch_effect(
  exp,
  group_col = "group",
  batch_col = "batch",
  prop_threshold = 0.3,
  check_confounding = TRUE,
  confounding_threshold = 0.4,
  info = NULL
)
}
\arguments{
\item{exp}{A \code{\link[glyexp:experiment]{glyexp::experiment()}} object.}

\item{group_col}{The column name in sample_info for groups. Default is "group".
Can be NULL when no group information is available.}

\item{batch_col}{The column name in sample_info for batches. Default is "batch".
Can be NULL when no batch information is available.}

\item{prop_threshold}{The proportion of variables that must show significant batch effects to perform batch correction.
Default is 0.3 (30\%).}

\item{check_confounding}{Whether to check for confounding between batch and group variables.
Default to TRUE.}

\item{confounding_threshold}{The threshold for Cramer's V to consider batch and group variables highly confounded.
Only used when \code{check_confounding} is TRUE. Default to 0.4.}

\item{info}{Internal parameter used by \code{\link[=auto_clean]{auto_clean()}}.}
}
\value{
A \code{\link[glyexp:experiment]{glyexp::experiment()}} object with batch effects corrected.
}
\description{
Detects and corrects batch effects in the experiment.
If batch information is available,
this function performs ANOVA to detect batch effects.
If more than 30\% (controlled by \code{prop_threshold}) of variables show significant batch effects (p < 0.05),
batch correction is performed using ComBat.
If group information exists,
it will be used as a covariate in both detection and correction
to preserve biological variation.
If no batch information is available,
the function will return the original experiment.
}
\examples{
exp <- glyexp::real_experiment
exp <- auto_batch_correction(exp)

}
