% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site-seq.R
\name{.validate_protein_seqs}
\alias{.validate_protein_seqs}
\title{Validate and normalize a named character vector of protein sequences}
\usage{
.validate_protein_seqs(seqs)
}
\description{
Validate and normalize a named character vector of protein sequences
}
\keyword{internal}
