% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{impute_bpca}
\alias{impute_bpca}
\title{BPCA Imputation}
\usage{
impute_bpca(x, by = NULL, ...)
}
\arguments{
\item{x}{A \code{\link[glyexp:experiment]{glyexp::experiment()}} object. \code{\link[glyexp:GlycomicSE]{glyexp::GlycomicSE()}} and
\code{\link[glyexp:GlycoproteomicSE]{glyexp::GlycoproteomicSE()}} objects are also supported.}

\item{by}{Either a column name in \code{sample_info} (string) or a factor/vector
specifying group assignments for each sample. Used for grouping when imputing missing values.}

\item{...}{Additional arguments to pass to \code{pcaMethods::pca()}.}
}
\value{
A \code{\link[glyexp:experiment]{glyexp::experiment()}} object with missing values imputed. Glyco
SE inputs return the same subclass.
}
\description{
A wrapper around the \code{\link[pcaMethods:pca]{pcaMethods::pca()}}.
Impute missing values using Bayesian principal component analysis (BPCA).
BPCA combines an EM approach for PCA with a Bayesian model.
In standard PCA data far from the training set but close to the principal
subspace may have the same reconstruction error.
BPCA defines a likelihood function such that the likelihood for data far from
the training set is much lower, even if they are close to the principal subspace.
}
