% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{impute_fw_knn}
\alias{impute_fw_knn}
\title{Feature-wise KNN Imputation}
\usage{
impute_fw_knn(x, k = 5, by = NULL, ...)
}
\arguments{
\item{x}{A \code{\link[glyexp:GlycomicSE]{glyexp::GlycomicSE()}}, \code{\link[glyexp:GlycoproteomicSE]{glyexp::GlycoproteomicSE()}}, or
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment::SummarizedExperiment()}} object.}

\item{k}{The number of nearest neighbors to consider.}

\item{by}{Either a column name in \code{sample_info} (string) or a factor/vector
specifying group assignments for each sample. Used for grouping when imputing missing values.}

\item{...}{Additional arguments to pass to \code{impute::impute.knn()}.}
}
\value{
A container of the same class as \code{x}, with missing values imputed.
}
\description{
A wrapper around the \code{\link[impute:impute.knn]{impute::impute.knn()}}.
Impute missing values with values from the k-nearest neighbors of the
corresponding feature.
}
