% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{impute_miss_forest}
\alias{impute_miss_forest}
\title{MissForest Imputation}
\usage{
impute_miss_forest(x, by = NULL, seed = 123, ...)
}
\arguments{
\item{x}{A supported glyco SE or legacy experiment container.}

\item{by}{Either a column name in \code{sample_info} (string) or a factor/vector
specifying group assignments for each sample. Used for grouping when imputing missing values.}

\item{seed}{Integer seed for random number generation. Default is 123.}

\item{...}{Additional arguments to pass to \code{missForest::missForest()}.}
}
\value{
The input container type with missing values imputed.
}
\description{
A wrapper around the \code{\link[missForest:missForest]{missForest::missForest()}}.
Impute missing values using recursive running of random forests until convergence.
This is a non-parametric method and works for both MAR and MNAR missing data.
}
