% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{impute_svd}
\alias{impute_svd}
\title{SVD Imputation}
\usage{
impute_svd(x, by = NULL, ...)
}
\arguments{
\item{x}{A \code{\link[glyexp:experiment]{glyexp::experiment()}} or
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment::SummarizedExperiment()}} object.}

\item{by}{Either a column name in \code{sample_info} (string) or a factor/vector
specifying group assignments for each sample. Used for grouping when imputing missing values.}

\item{...}{Additional arguments to pass to \code{pcaMethods::pca()}.}
}
\value{
A \code{\link[glyexp:experiment]{glyexp::experiment()}} object with missing values imputed. SummarizedExperiment
inputs return the same class.
}
\description{
A wrapper around the \code{\link[pcaMethods:pca]{pcaMethods::pca()}}.
Impute missing values using singular value decomposition (SVD) imputation.
SVD is a matrix factorization technique that factors a matrix into three matrices:
U, Σ, and V. SVD is used to find the best lower rank approximation of the original matrix.
}
