% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{impute_sw_knn}
\alias{impute_sw_knn}
\title{Sample-wise KNN Imputation}
\usage{
impute_sw_knn(x, k = 5, by = NULL, ...)
}
\arguments{
\item{x}{A supported glyco SE or legacy experiment container.}

\item{k}{The number of nearest neighbors to consider.}

\item{by}{Either a column name in \code{sample_info} (string) or a factor/vector
specifying group assignments for each sample. Used for grouping when imputing missing values.}

\item{...}{Additional arguments to pass to \code{impute::impute.knn()}.}
}
\value{
The input container type with missing values imputed.
}
\description{
A wrapper around the \code{\link[impute:impute.knn]{impute::impute.knn()}}.
Impute missing values with values from the k-nearest neighbors of the
corresponding sample.
If there are strong patterns among the samples (such as group clustering
relationships or experimental conditions), this method can better utilize
the overall relationships among samples.
}
