% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{median_abs_normalize}
\alias{median_abs_normalize}
\title{Absolute Median Normalization}
\usage{
median_abs_normalize(exp)
}
\arguments{
\item{exp}{An experiment object.}
}
\value{
An experiment object with the expression matrix normalized.
}
\description{
Normalize the expression matrix by dividing each column (sample) by the absolute
median of that column (NA ignored), so that the absolute median of each column is 1.
}
