% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{min_prob_impute}
\alias{min_prob_impute}
\title{Minimum Probability Imputation}
\usage{
min_prob_impute(exp, by = NULL, ...)
}
\arguments{
\item{exp}{An expression matrix.}

\item{by}{A grouping variable to consider when imputing missing values.
This variable should be a column in the sample information table.}

\item{...}{Additional arguments to pass to \code{imputeLCMD::impute.MinProb()}.}
}
\value{
An expression matrix with missing values imputed.
}
\description{
A wrapper around the \code{\link[imputeLCMD:impute.MinProb]{imputeLCMD::impute.MinProb()}}.
Impute missing values using random draws from the left-censored
gaussian distribution.
}
