% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize_loesscyc}
\alias{normalize_loesscyc}
\title{LoessCyc Normalization}
\usage{
normalize_loesscyc(x, by = NULL, ...)
}
\arguments{
\item{x}{A supported glyco SE or legacy experiment container.}

\item{by}{Either a column name in \code{sample_info} (string) or a factor/vector
specifying group assignments for each sample. Optional.
If provided, the normalization will be performed within each group.}

\item{...}{Additional arguments to pass to \code{\link[limma:normalizeCyclicLoess]{limma::normalizeCyclicLoess()}}.}
}
\value{
The input container type with a normalized expression matrix.
}
\description{
This function is a wrapper around \code{\link[limma:normalizeCyclicLoess]{limma::normalizeCyclicLoess()}} with
\code{method = "pairs"}.
Each pair of columns is normalized mutually to each other.
See \href{https://doi.org/10.1093/bib/bbw095}{this paper} for more information.
Also see \code{\link[limma:normalizeCyclicLoess]{limma::normalizeCyclicLoess()}}.
}
