% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize_quantile}
\alias{normalize_quantile}
\title{Quantile Normalization}
\usage{
normalize_quantile(x, by = NULL, ...)
}
\arguments{
\item{x}{A \code{\link[glyexp:experiment]{glyexp::experiment()}} object.}

\item{by}{Either a column name in \code{sample_info} (string) or a factor/vector
specifying group assignments for each sample. Optional.
If provided, the normalization will be performed within each group.}

\item{...}{Additional arguments to pass to \code{\link[limma:normalizequantiles]{limma::normalizeQuantiles()}}.}
}
\value{
A \code{\link[glyexp:experiment]{glyexp::experiment()}} object with a normalized expression matrix.
}
\description{
This function is a wrapper around \code{\link[limma:normalizequantiles]{limma::normalizeQuantiles()}}.
It normalizes the expression matrix by quantile normalization.
This method is used to remove technical variation between samples.
Proteomics data rarely uses this method, but it is common in microarray data.
See \href{https://en.wikipedia.org/wiki/Quantile_normalization}{wikipedia}
for more information.
}
