% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize_rlrma}
\alias{normalize_rlrma}
\title{Robust Linear Regression with Median Adjustment Normalization}
\usage{
normalize_rlrma(x, by = NULL)
}
\arguments{
\item{x}{A \code{\link[glyexp:GlycomicSE]{glyexp::GlycomicSE()}}, \code{\link[glyexp:GlycoproteomicSE]{glyexp::GlycoproteomicSE()}}, or
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment::SummarizedExperiment()}} object.}

\item{by}{Either a column name in \code{sample_info} (string) or a factor/vector
specifying group assignments for each sample. Optional.
If provided, the normalization will be performed within each group.}
}
\value{
A container of the same class as \code{x}, with a normalized expression
matrix.
}
\description{
This method is based on robust linear regression with median adjustment.
First, the median of each variable's abundance across all measured samples
is subtracted from the sample's abundance values. Then, it's like the
\code{\link[=normalize_rlr]{normalize_rlr()}} method.
}
