% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize_total_area}
\alias{normalize_total_area}
\title{Total Area Normalization}
\usage{
normalize_total_area(x)
}
\arguments{
\item{x}{A \code{\link[glyexp:experiment]{glyexp::experiment()}} or
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment::SummarizedExperiment()}} object.}
}
\value{
A \code{\link[glyexp:experiment]{glyexp::experiment()}} object with a normalized expression matrix.
SummarizedExperiment inputs return the same class.
}
\description{
Normalize the expression matrix by dividing each column (sample) by the sum of
that column, so that the sum of each column is 1.
This is the most common normalization method for glycomics data.
It removes the bias introduced by total glycan abundance.
However, it results in compositional data, which may result in unrealistic
downstream analysis results.
}
