% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{plot_batch_pca}
\alias{plot_batch_pca}
\title{Plot PCA Score by Batch}
\usage{
plot_batch_pca(exp, batch_col = "batch")
}
\arguments{
\item{exp}{A \code{\link[glyexp:experiment]{glyexp::experiment()}} or
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment::SummarizedExperiment()}} object.}

\item{batch_col}{Column name in \code{sample_info}, or a factor/vector with length
equal to the number of samples.}
}
\value{
A ggplot object of PCA scores.
}
\description{
Draw a PCA score plot for samples and color points by batch.
PCA is computed on log2-transformed intensities after removing variables
with missing values.
}
\examples{
library(SummarizedExperiment)

exp <- glyexp::real_experiment
batch <- rep(c("A", "B"), length.out = ncol(exp))
if (inherits(exp, "glyexp_experiment")) {
  exp$sample_info$batch <- batch
} else {
  colData(exp)$batch <- batch
}
plot_batch_pca(exp, batch_col = "batch")

}
