% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{plot_cv_dent}
\alias{plot_cv_dent}
\title{Plot CV Density}
\usage{
plot_cv_dent(exp, by = NULL)
}
\arguments{
\item{exp}{A \code{\link[glyexp:GlycomicSE]{glyexp::GlycomicSE()}}, \code{\link[glyexp:GlycoproteomicSE]{glyexp::GlycoproteomicSE()}}, or
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment::SummarizedExperiment()}} object.}

\item{by}{Grouping variable for samples. Can be a column name in \code{sample_info}
or a vector/factor with length equal to the number of samples. When provided,
CVs are computed within each group and densities are shown with different fills.}
}
\value{
A ggplot object of CV density.
}
\description{
Compute coefficient of variation (CV) for each variable and plot its density.
When \code{by} is provided, CVs are computed within each group and densities are
shown with different fills.
}
\examples{
plot_cv_dent(glyexp::real_experiment)
library(SummarizedExperiment)

exp <- glyexp::real_experiment
group <- rep(c("A", "B"), length.out = ncol(exp))
colData(exp)$group <- group
plot_cv_dent(exp, by = "group")

}
