% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{plot_int_boxplot}
\alias{plot_int_boxplot}
\title{Plot Log-Intensity Boxplots by Sample}
\usage{
plot_int_boxplot(exp, by = NULL)
}
\arguments{
\item{exp}{A \code{\link[glyexp:experiment]{glyexp::experiment()}} object. \code{\link[glyexp:GlycomicSE]{glyexp::GlycomicSE()}} and
\code{\link[glyexp:GlycoproteomicSE]{glyexp::GlycoproteomicSE()}} objects are also supported.}

\item{by}{Grouping variable for samples. Can be a column name in \code{sample_info}
or a vector/factor with length equal to the number of samples. When provided,
samples are grouped along the x-axis and boxplots are colored by group.}
}
\value{
A ggplot object of log-intensity boxplots.
}
\description{
Draw boxplots of log2-transformed intensities for each sample.
Optionally color and group samples by a metadata variable.
}
\examples{
plot_int_boxplot(glyexp::toy_experiment)
plot_int_boxplot(glyexp::toy_experiment, by = "group")

}
