% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{plot_missing_bar}
\alias{plot_missing_bar}
\title{Plot Missing Value Proportions by Sample}
\usage{
plot_missing_bar(exp, ...)
}
\arguments{
\item{exp}{A \code{\link[glyexp:experiment]{glyexp::experiment()}} object.}

\item{...}{Other arguments passed to \code{ggplot2::geom_col()}.}
}
\value{
A ggplot object of missing value proportions by sample.
}
\description{
Draw a bar plot of missing value proportions for each sample.
Samples are ordered from low to high missing proportion.
}
\examples{
plot_missing_bar(glyexp::toy_experiment)

}
