% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{plot_missing_heatmap}
\alias{plot_missing_heatmap}
\title{Plot Missing Value Heatmap}
\usage{
plot_missing_heatmap(exp, ...)
}
\arguments{
\item{exp}{A \code{\link[glyexp:experiment]{glyexp::experiment()}} object.}

\item{...}{Other arguments passed to \code{pheatmap::pheatmap()}.}
}
\value{
A ggplot object of the missing value heatmap.
}
\description{
Draw a binary heatmap showing the missing value pattern in an experiment.
Values are binarized: 1 (present) or 0 (missing).
Rows (variables) are sorted by missing value proportion from low to high.
Columns (samples) are clustered using hierarchical clustering.
}
\examples{
plot_missing_heatmap(glyexp::toy_experiment)

}
