% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{plot_rank_abundance}
\alias{plot_rank_abundance}
\title{Plot Rank Abundance}
\usage{
plot_rank_abundance(exp)
}
\arguments{
\item{exp}{A supported glyco SE or legacy experiment container.}
}
\value{
A ggplot object of protein rank abundance.
}
\description{
Draw a scatter plot of proteins ranked by mean log2 intensity.
Proteins are ordered from high to low mean intensity along the x-axis.
}
\examples{
plot_rank_abundance(glyexp::toy_experiment)

}
