% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{plot_rle}
\alias{plot_rle}
\title{Plot Relative Log Expression (RLE) Boxplots}
\usage{
plot_rle(exp, by = NULL)
}
\arguments{
\item{exp}{A \code{\link[glyexp:experiment]{glyexp::experiment()}} object.}

\item{by}{Grouping variable for samples. Can be a column name in \code{sample_info}
or a vector/factor with length equal to the number of samples. When provided,
samples are grouped along the x-axis and boxplots are colored by group.}
}
\value{
A ggplot object of RLE boxplots.
}
\description{
Draw boxplots of relative log expression (log2 intensity minus row median)
for each sample. Optionally color and group samples by a metadata variable.
}
\examples{
plot_rle(glyexp::toy_experiment)
plot_rle(glyexp::toy_experiment, by = "group")

}
