% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{ppca_impute}
\alias{ppca_impute}
\title{PPCA Imputation}
\usage{
ppca_impute(exp, by = NULL, ...)
}
\arguments{
\item{exp}{An expression matrix.}

\item{by}{A grouping variable to consider when imputing missing values.
This variable should be a column in the sample information table.}

\item{...}{Additional arguments to pass to \code{pcaMethods::pca()}.}
}
\value{
An expression matrix with missing values imputed.
}
\description{
A wrapper around the \code{\link[pcaMethods:pca]{pcaMethods::pca()}}.
Impute missing values using probabilistic principal component analysis (PPCA).
PPCA allows to perform PCA on incomplete data and may be used for missing value estimation.
}
