% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{quantile_normalize}
\alias{quantile_normalize}
\title{Quantile Normalization}
\usage{
quantile_normalize(exp)
}
\arguments{
\item{exp}{An experiment object.}
}
\value{
An experiment object with the expression matrix normalized.
}
\description{
This function is a wrapper around \code{\link[limma:normalizequantiles]{limma::normalizeQuantiles()}}.
It normalizes the expression matrix by quantile normalization.
This method is used to remove technical variation between samples.
Proteomics data rarely uses this method, but it is common in microarray data.
See \href{https://en.wikipedia.org/wiki/Quantile_normalization}{wikipedia}
for more information.
}
