% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.R
\name{remove_constant}
\alias{remove_constant}
\title{Remove Constant Variables}
\usage{
remove_constant(x, by = NULL, strict = FALSE)
}
\arguments{
\item{x}{Either a \code{glyexp_experiment} object or a matrix.}

\item{by}{Either a column name in \code{sample_info} (string) or a vector specifying group assignments for each sample.}

\item{strict}{If \code{FALSE}, remove a variable only if it is constant in all groups.
If \code{TRUE}, remove a variable if it is constant in any group. Defaults to FALSE.}
}
\value{
For \code{glyexp_experiment} input, returns a modified \code{glyexp_experiment} object.
For matrix input, returns a filtered matrix.
}
\description{
Constant variables are variables with the same value in all samples.
This function is equivalent to \code{remove_low_var(x, var_cutoff = 0, by = by, strict = strict)}.
}
\seealso{
\code{\link[=remove_low_var]{remove_low_var()}}
}
