% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.R
\name{remove_low_expr}
\alias{remove_low_expr}
\title{Remove Variables with Low Expression}
\usage{
remove_low_expr(x, percentile = 0.05, by = NULL, strict = FALSE)
}
\arguments{
\item{x}{A \code{\link[glyexp:experiment]{glyexp::experiment()}} or
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment::SummarizedExperiment()}} object.}

\item{percentile}{The percentile for median expression values.
Defaults to 0.05, i.e., the 5\% lowest median expression values will be removed.}

\item{by}{Either a column name in \code{sample_info} (string) or a vector specifying group assignments for each sample.}

\item{strict}{If \code{FALSE}, remove a variable only if it passes the abundance thresholds in all groups.
If \code{TRUE}, remove a variable if it passes the abundance thresholds in any group. Defaults to FALSE.}
}
\value{
A \code{\link[glyexp:experiment]{glyexp::experiment()}} object with filtered variables. SummarizedExperiment inputs
return the same class.
}
\description{
Filters variables based on median expression values.
Variables with median expression values below certain percentile will be removed.
}
