% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.R
\name{remove_low_var}
\alias{remove_low_var}
\title{Remove Variables with Low Variance
Remove Variables with Low Variance}
\usage{
remove_low_var(x, var_cutoff = 0, by = NULL, strict = FALSE)
}
\arguments{
\item{x}{A \code{\link[glyexp:experiment]{glyexp::experiment()}} or
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment::SummarizedExperiment()}} object.}

\item{var_cutoff}{The cutoff for variance. Defaults to 0.}

\item{by}{A factor specifying the groupings. Defaults to NULL.}

\item{strict}{If \code{FALSE}, remove a variable only if it passes the variance threshold in all groups.
If \code{TRUE}, remove a variable if it passes the variance threshold in any group.}
}
\value{
A \code{\link[glyexp:experiment]{glyexp::experiment()}} object with filtered variables. SummarizedExperiment inputs
return the same class.
}
\description{
Filters variables whose variance falls below a threshold.
Default behavior is to remove variables with zero variance.
}
\seealso{
\code{\link[=remove_low_cv]{remove_low_cv()}}, \code{\link[=remove_constant]{remove_constant()}}
}
