% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{rlr_normalize}
\alias{rlr_normalize}
\title{Robust Linear Regression Normalization}
\usage{
rlr_normalize(exp, by = NULL)
}
\arguments{
\item{exp}{An experiment object.}

\item{by}{A column name in \code{sample_info} to stratify by. Optional.
If provided, the normalization will be performed within each group.}
}
\value{
An experiment object with the expression matrix normalized.
}
\description{
This method is based on robust linear regression. The reference sample is
calculated as the median value of each variable's abundance across all
measured samples. For each sample, a robust linear regression model is
fitted to the sample's abundance values against the reference sample's
abundance values. The fitted model is then used to normalize the sample's
abundance values. The underlying assumption is that the diﬀerent intensities
observed across individuals are imputable to diﬀerent amounts of the biological
material in the collected samples.
}
