% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{rlrmacyc_normalize}
\alias{rlrmacyc_normalize}
\title{Robust Linear Regression with Median Adjustment and Cyclic Normalization}
\usage{
rlrmacyc_normalize(exp, n_iter = 3, by = NULL)
}
\arguments{
\item{exp}{An experiment object.}

\item{n_iter}{The number of iterations to perform. Default is 3.}

\item{by}{A column name in \code{sample_info} to stratify by. Optional.
If provided, the normalization will be performed within each group.}
}
\value{
An experiment object with the expression matrix normalized.
}
\description{
This method is based on robust linear regression with median adjustment
and cyclic normalization. The method is applied iteratively to each pair
of samples. For each pair of samples, the median of the differences between
the two samples is calculated. A robust linear regression model is fitted
to the differences against the averages of the two samples. The fitted model
is then used to normalize the two samples. The process is repeated for a
number of iterations.
}
