% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{ta_normalize}
\alias{ta_normalize}
\title{Total Abundance Normalization}
\usage{
ta_normalize(exp)
}
\arguments{
\item{exp}{An experiment object.}
}
\value{
An experiment object with the expression matrix normalized.
}
\description{
Normalize the expression matrix by dividing each column (sample) by the sum of
that column, so that the sum of each column is 1.
This is the most common normalization method for glycomics data.
It removes the bias introduced by total glycan abundance.
However, it results in compositional data, which may result in unrealistic
downstream analysis results.
}
