% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{zero_impute}
\alias{zero_impute}
\title{Zero Imputation}
\usage{
zero_impute(exp)
}
\arguments{
\item{exp}{An expression matrix.}
}
\value{
An expression matrix with missing values imputed.
}
\description{
Impute missing values in an expression matrix by replacing them with zeros.
}
