% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{aggregate}
\alias{aggregate}
\title{Aggregate Data}
\usage{
aggregate(exp, to_level = c("gf", "gp", "gfs", "gps"))
}
\arguments{
\item{exp}{A \code{glyexp_experiment} object containing glycoproteomics data.
This function only works with \code{glyexp_experiment} objects as it requires
variable information for aggregation.}

\item{to_level}{The aggregation level,
one of: "gf" (glycoforms), "gp" (glycopeptides),
"gfs" (glycoforms with structures),
or "gps" (glycopeptides with structures).
See Details for more information.}
}
\value{
A modified \code{glyexp_experiment} object
with aggregated expression matrix and
updated variable information.
}
\description{
Aggregate glycoproteomics data to different levels
(glycoforms, glycopeptides, etc.).
This function sums up quantitative values for each
unique combination of specified variables.
It is recommended to call this function after missing value imputation.

The following levels are available:
\itemize{
\item "gf": Aggregate to glycoforms, which is the unique combination of proteins,
protein sites, and glycan compositions.
This is the default level.
\item "gp": Aggregate to glycopeptides,
which is the unique combination of peptides,
proteins, protein sites, and glycan compositions.
\item "gfs": Like "gf", but differentiates structures with the same composition.
\item "gps": Like "gp", but differentiates structures with the same composition.
}
}
\details{
Different levels of aggregation require different columns in the variable information.
\itemize{
\item "gf": "protein", "gene", "glycan_composition", "protein_site"
\item "gp": "peptide", "protein", "gene", "glycan_composition", "peptide_site", "protein_site"
\item "gfs": "protein", "gene", "glycan_composition", "glycan_structure", "protein_site"
\item "gps": "peptide", "protein", "gene", "glycan_composition", "glycan_structure",
"peptide_site", "protein_site"
}
}
