% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{aggregate}
\alias{aggregate}
\alias{glyclean_aggregate.glyexp_experiment}
\alias{glyclean_aggregate.default}
\title{Aggregate Data}
\usage{
aggregate(
  exp,
  to_level = c("gf", "gp", "gfs", "gps"),
  standardize_variable = TRUE
)

\method{glyclean_aggregate}{glyexp_experiment}(
  exp,
  to_level = c("gf", "gp", "gfs", "gps"),
  standardize_variable = TRUE
)

\method{glyclean_aggregate}{default}(
  exp,
  to_level = c("gf", "gp", "gfs", "gps"),
  standardize_variable = TRUE
)
}
\arguments{
\item{exp}{A \code{\link[glyexp:experiment]{glyexp::experiment()}} object with "glycoproteomics" type.}

\item{to_level}{The aggregation level,
one of: "gf" (glycoforms), "gp" (glycopeptides),
"gfs" (glycoforms with structures),
or "gps" (glycopeptides with structures).
See Details for more information.}

\item{standardize_variable}{Whether to call \code{\link[glyexp:standardize_variable]{glyexp::standardize_variable()}}
after aggregation. Set to \code{FALSE} to skip network calls for faster testing.
Default is \code{TRUE}.}
}
\value{
A modified \code{\link[glyexp:experiment]{glyexp::experiment()}} object with aggregated expression matrix and
updated variable information.
}
\description{
Aggregate glycoproteomics data to different levels
(glycoforms, glycopeptides, etc.).
This function sums up quantitative values for each
unique combination of specified variables.
It is recommended to call this function after missing value imputation.

The following levels are available:
\itemize{
\item "gf": Aggregate to glycoforms, which is the unique combination of proteins,
protein sites, and glycan compositions.
This is the default level.
\item "gp": Aggregate to glycopeptides,
which is the unique combination of peptides,
proteins, protein sites, and glycan compositions.
\item "gfs": Like "gf", but differentiates structures with the same composition.
\item "gps": Like "gp", but differentiates structures with the same composition.
}

Different levels of aggregation require different columns in the variable information.
\itemize{
\item "gf": "protein", "glycan_composition", "protein_site"
\item "gp": "peptide", "protein", "glycan_composition", "peptide_site", "protein_site"
\item "gfs": "protein", "glycan_composition", "glycan_structure", "protein_site"
\item "gps": "peptide", "protein", "glycan_composition", "glycan_structure",
"peptide_site", "protein_site"
}

Other columns in the variable information tibble with "many-to-one" relationship
with the unique combination of columns listed above will be kept.
A common example is the "gene" column.
For each "glycoform" (unique combination of "protein", "protein_site", and "glycan_composition"),
there should be only one "gene" value, therefore it is kept for "gf" level.
On the other hand, the "peptide" column is removed for "gf" level,
as one "glycoform" can contain multiple "peptides".
}
