% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto-aggregate.R
\name{auto_aggregate}
\alias{auto_aggregate}
\title{Automatic Aggregation}
\usage{
auto_aggregate(exp)
}
\arguments{
\item{exp}{A \code{\link[glyexp:experiment]{glyexp::experiment()}} object with "glycoproteomics" type.}
}
\value{
A modified \code{\link[glyexp:experiment]{glyexp::experiment()}} object with aggregated expression matrix and
updated variable information.
}
\description{
Aggregates glycoproteomics data to "gfs" (glycoforms with structures) level
if the glycan structure column exists,
otherwise to "gf" (glycoforms with compositions) level.
}
\examples{
library(glyexp)
exp <- real_experiment
auto_aggregate(exp)

}
