% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto-remove.R
\name{auto_remove}
\alias{auto_remove}
\title{Automatic Removing Variables}
\usage{
auto_remove(
  exp,
  preset = "discovery",
  group_col = "group",
  qc_name = "QC",
  info = NULL
)
}
\arguments{
\item{exp}{A glyexp_experiment object.}

\item{preset}{One of "simple", "discovery", or "biomarker".
Default "discovery" if group information is available, otherwise "simple".}

\item{group_col}{The column name in sample_info for groups. Default is "group".
Can be NULL when no group information is available.}

\item{qc_name}{The name of QC samples in the \code{group_col} column. Default is "QC".
Only used when \code{group_col} is not NULL. Can be NULL when no QC samples are available.}

\item{info}{Internal parameter used by \code{\link[=auto_clean]{auto_clean()}}.}
}
\value{
A modified \code{\link[glyexp:experiment]{glyexp::experiment()}} object.
}
\description{
This function uses preset rules to remove variables with low quality.
Available presets:
\itemize{
\item "simple": remove variables with more than 50\% missing values.
\item "discovery": more lenient, remove variables with more than 80\% missing values,
but ensure less than 50\% of missing values in at least one group.
\item "biomarker": more strict, remove variables with more than 40\% missing values,
and ensure less than 60\% of missing values in all groups.
}

QC samples will not be considered in the removal process.
}
\examples{
library(glyexp)
exp <- real_experiment
auto_remove(exp)

}
