% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch-correction.R
\name{detect_batch_effect}
\alias{detect_batch_effect}
\title{Detect batch effect}
\usage{
detect_batch_effect(x, batch = "batch", group = NULL)
}
\arguments{
\item{x}{A \code{\link[glyexp:experiment]{glyexp::experiment()}} or
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment::SummarizedExperiment()}} object.}

\item{batch}{Either a factor/character vector specifying batch assignments for each sample,
or a string specifying the column name in \code{sample_info}. Defaults to "batch".}

\item{group}{Either a factor/character vector specifying group assignments for each sample,
or a string specifying the column name in \code{sample_info}.
If provided, it will be used as a covariate in the ANOVA model.
This is useful when you have an unbalanced design.
Default to NULL.}
}
\value{
A double vector of p-values for each variable, with the same length
as \code{nrow(glyexp::get_expr_mat(x))}.
}
\description{
Use ANOVA to detect if batch effect is present in the data.
If \code{group} is provided, it will be used as a covariate in the ANOVA model.
}
\examples{
library(SummarizedExperiment)

# With a SummarizedExperiment and column names
exp <- glyexp::real_experiment
batch <- rep(c("A", "B"), length.out = ncol(exp))
group <- rep(c("Ctrl", "Ctrl", "Treat", "Treat"), length.out = ncol(exp))
if (inherits(exp, "glyexp_experiment")) {
  exp$sample_info$batch <- batch
  exp$sample_info$group <- group
} else {
  SummarizedExperiment::colData(exp)$batch <- batch
  SummarizedExperiment::colData(exp)$group <- group
}
p_values <- detect_batch_effect(exp, batch = "batch", group = "group")

}
