% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{impute_bpca}
\alias{impute_bpca}
\title{BPCA Imputation}
\usage{
impute_bpca(x, by = NULL, ...)
}
\arguments{
\item{x}{Either a \code{glyexp_experiment} object or a matrix.
If a matrix, rows should be variables and columns should be samples.}

\item{by}{Either a column name in \code{sample_info} (string) or a factor/vector
specifying group assignments for each sample. Used for grouping when imputing missing values.}

\item{...}{Additional arguments to pass to \code{pcaMethods::pca()}.}
}
\value{
Returns the same type as the input. If \code{x} is a \code{glyexp_experiment},
returns a \code{glyexp_experiment} with missing values imputed.
If \code{x} is a matrix, returns a matrix with missing values imputed.
}
\description{
A wrapper around the \code{\link[pcaMethods:pca]{pcaMethods::pca()}}.
Impute missing values using Bayesian principal component analysis (BPCA).
BPCA combines an EM approach for PCA with a Bayesian model.
In standard PCA data far from the training set but close to the principal
subspace may have the same reconstruction error.
BPCA defines a likelihood function such that the likelihood for data far from
the training set is much lower, even if they are close to the principal subspace.
}
