% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{impute_bpca}
\alias{impute_bpca}
\title{BPCA Imputation}
\usage{
impute_bpca(exp, by = NULL, ...)
}
\arguments{
\item{exp}{An expression matrix.}

\item{by}{A grouping variable to consider when imputing missing values.
This variable should be a column in the sample information table.}

\item{...}{Additional arguments to pass to \code{pcaMethods::pca()}.}
}
\value{
An expression matrix with missing values imputed.
}
\description{
A wrapper around the \code{\link[pcaMethods:pca]{pcaMethods::pca()}}.
Impute missing values using Bayesian principal component analysis (BPCA).
BPCA combines an EM approach for PCA with a Bayesian model.
In standard PCA data far from the training set but close to the principal
subspace may have the same reconstruction error.
BPCA defines a likelihood function such that the likelihood for data far from
the training set is much lower, even if they are close to the principal subspace.
}
