% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{impute_fw_knn}
\alias{impute_fw_knn}
\title{Feature-wise KNN Imputation}
\usage{
impute_fw_knn(x, k = 5, by = NULL, ...)
}
\arguments{
\item{x}{Either a \code{glyexp_experiment} object or a matrix.
If a matrix, rows should be variables and columns should be samples.}

\item{k}{The number of nearest neighbors to consider.}

\item{by}{A grouping variable to consider when imputing missing values.
This variable should be a column in the variable information table.
Only available for \code{glyexp_experiment} objects.}

\item{...}{Additional arguments to pass to \code{impute::impute.knn()}.}
}
\value{
Returns the same type as the input. If \code{x} is a \code{glyexp_experiment},
returns a \code{glyexp_experiment} with missing values imputed.
If \code{x} is a matrix, returns a matrix with missing values imputed.
}
\description{
A wrapper around the \code{\link[impute:impute.knn]{impute::impute.knn()}}.
Impute missing values with values from the k-nearest neighbors of the
corresponding feature.
}
