% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{impute_half_sample_min}
\alias{impute_half_sample_min}
\alias{impute_half_sample_min.glyexp_experiment}
\alias{impute_half_sample_min.matrix}
\alias{impute_half_sample_min.default}
\title{Half Sample Minimum Imputation}
\usage{
impute_half_sample_min(x)

\method{impute_half_sample_min}{glyexp_experiment}(x)

\method{impute_half_sample_min}{matrix}(x)

\method{impute_half_sample_min}{default}(x)
}
\arguments{
\item{x}{Either a \code{glyexp_experiment} object or a matrix.
If a matrix, rows should be variables and columns should be samples.}
}
\value{
Returns the same type as the input. If \code{x} is a \code{glyexp_experiment},
returns a \code{glyexp_experiment} with missing values imputed.
If \code{x} is a matrix, returns a matrix with missing values imputed.
}
\description{
Impute missing values with half of the minimum value of the corresponding sample.
This method assumes that missing values are MCAR,
i.e. missing values are induced by an ion below the detection limit.
Compared to \code{\link[=impute_sample_min]{impute_sample_min()}}, this method is more conservative.
}
