% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{impute_half_sample_min}
\alias{impute_half_sample_min}
\title{Half Sample Minimum Imputation}
\usage{
impute_half_sample_min(exp)
}
\arguments{
\item{exp}{An expression matrix.}
}
\value{
An expression matrix with missing values imputed.
}
\description{
Impute missing values with half of the minimum value of the corresponding sample.
This method assumes that missing values are MCAR,
i.e. missing values are induced by an ion below the detection limit.
Compared to \code{\link[=impute_sample_min]{impute_sample_min()}}, this method is more conservative.
}
