% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{impute_min_prob}
\alias{impute_min_prob}
\title{Minimum Probability Imputation}
\usage{
impute_min_prob(x, by = NULL, ...)
}
\arguments{
\item{x}{Either a \code{glyexp_experiment} object or a matrix.
If a matrix, rows should be variables and columns should be samples.}

\item{by}{A grouping variable to consider when imputing missing values.
This variable should be a column in the sample information table.
Only available for \code{glyexp_experiment} objects.}

\item{...}{Additional arguments to pass to \code{imputeLCMD::impute.MinProb()}.}
}
\value{
Returns the same type as the input. If \code{x} is a \code{glyexp_experiment},
returns a \code{glyexp_experiment} with missing values imputed.
If \code{x} is a matrix, returns a matrix with missing values imputed.
}
\description{
A wrapper around the \code{\link[imputeLCMD:impute.MinProb]{imputeLCMD::impute.MinProb()}}.
Impute missing values using random draws from the left-censored
gaussian distribution.
}
