% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{impute_miss_forest}
\alias{impute_miss_forest}
\title{MissForest Imputation}
\usage{
impute_miss_forest(exp, by = NULL, ...)
}
\arguments{
\item{exp}{An expression matrix.}

\item{by}{A grouping variable to consider when imputing missing values.
This variable should be a column in the sample information table.}

\item{...}{Additional arguments to pass to \code{missForest::missForest()}.}
}
\value{
An expression matrix with missing values imputed.
}
\description{
A wrapper around the \code{\link[missForest:missForest]{missForest::missForest()}}.
Impute missing values using recursive running of random forests until convergence.
This is a non-parametric method and works for both MAR and MNAR missing data.
}
