% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{impute_miss_forest}
\alias{impute_miss_forest}
\title{MissForest Imputation}
\usage{
impute_miss_forest(x, by = NULL, ...)
}
\arguments{
\item{x}{Either a \code{glyexp_experiment} object or a matrix.
If a matrix, rows should be variables and columns should be samples.}

\item{by}{Either a column name in \code{sample_info} (string) or a factor/vector
specifying group assignments for each sample. Used for grouping when imputing missing values.}

\item{...}{Additional arguments to pass to \code{missForest::missForest()}.}
}
\value{
Returns the same type as the input. If \code{x} is a \code{glyexp_experiment},
returns a \code{glyexp_experiment} with missing values imputed.
If \code{x} is a matrix, returns a matrix with missing values imputed.
}
\description{
A wrapper around the \code{\link[missForest:missForest]{missForest::missForest()}}.
Impute missing values using recursive running of random forests until convergence.
This is a non-parametric method and works for both MAR and MNAR missing data.
}
