% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{impute_ppca}
\alias{impute_ppca}
\title{PPCA Imputation}
\usage{
impute_ppca(x, by = NULL, ..., seed = 123)
}
\arguments{
\item{x}{A \code{\link[glyexp:GlycomicSE]{glyexp::GlycomicSE()}}, \code{\link[glyexp:GlycoproteomicSE]{glyexp::GlycoproteomicSE()}}, or
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment::SummarizedExperiment()}} object.}

\item{by}{Either a column name in \code{sample_info} (string) or a factor/vector
specifying group assignments for each sample. Used for grouping when imputing missing values.}

\item{...}{Additional arguments to pass to \code{pcaMethods::pca()}.}

\item{seed}{Integer seed for random number generation. Default is \code{123}.}
}
\value{
A container of the same class as \code{x}, with missing values imputed.
}
\description{
A wrapper around the \code{\link[pcaMethods:pca]{pcaMethods::pca()}}.
Impute missing values using probabilistic principal component analysis (PPCA).
PPCA allows to perform PCA on incomplete data and may be used for missing value estimation.
}
