% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{impute_ppca}
\alias{impute_ppca}
\title{PPCA Imputation}
\usage{
impute_ppca(x, by = NULL, ...)
}
\arguments{
\item{x}{Either a \code{glyexp_experiment} object or a matrix.
If a matrix, rows should be variables and columns should be samples.}

\item{by}{Either a column name in \code{sample_info} (string) or a factor/vector
specifying group assignments for each sample. Used for grouping when imputing missing values.}

\item{...}{Additional arguments to pass to \code{pcaMethods::pca()}.}
}
\value{
Returns the same type as the input. If \code{x} is a \code{glyexp_experiment},
returns a \code{glyexp_experiment} with missing values imputed.
If \code{x} is a matrix, returns a matrix with missing values imputed.
}
\description{
A wrapper around the \code{\link[pcaMethods:pca]{pcaMethods::pca()}}.
Impute missing values using probabilistic principal component analysis (PPCA).
PPCA allows to perform PCA on incomplete data and may be used for missing value estimation.
}
