% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{impute_sample_min}
\alias{impute_sample_min}
\alias{impute_sample_min.glyexp_experiment}
\alias{impute_sample_min.matrix}
\alias{impute_sample_min.default}
\title{Sample Minimum Imputation}
\usage{
impute_sample_min(x)

\method{impute_sample_min}{glyexp_experiment}(x)

\method{impute_sample_min}{matrix}(x)

\method{impute_sample_min}{default}(x)
}
\arguments{
\item{x}{Either a \code{glyexp_experiment} object or a matrix.
If a matrix, rows should be variables and columns should be samples.}
}
\value{
Returns the same type as the input. If \code{x} is a \code{glyexp_experiment},
returns a \code{glyexp_experiment} with missing values imputed.
If \code{x} is a matrix, returns a matrix with missing values imputed.
}
\description{
Impute missing values with the minimum value of the corresponding sample.
This method assumes that missing values are MCAR,
i.e. missing values are induced by an ion below the detection limit.
See also \code{\link[=impute_half_sample_min]{impute_half_sample_min()}}.
}
