% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{impute_svd}
\alias{impute_svd}
\alias{impute_svd.glyexp_experiment}
\alias{impute_svd.matrix}
\alias{impute_svd.default}
\title{SVD Imputation}
\usage{
impute_svd(x, by = NULL, ...)

\method{impute_svd}{glyexp_experiment}(x, by = NULL, ...)

\method{impute_svd}{matrix}(x, by = NULL, ...)

\method{impute_svd}{default}(x, by = NULL, ...)
}
\arguments{
\item{x}{Either a \code{glyexp_experiment} object or a matrix.
If a matrix, rows should be variables and columns should be samples.}

\item{by}{Either a column name in \code{sample_info} (string) or a factor/vector
specifying group assignments for each sample. Used for grouping when imputing missing values.}

\item{...}{Additional arguments to pass to \code{pcaMethods::pca()}.}
}
\value{
Returns the same type as the input. If \code{x} is a \code{glyexp_experiment},
returns a \code{glyexp_experiment} with missing values imputed.
If \code{x} is a matrix, returns a matrix with missing values imputed.
}
\description{
A wrapper around the \code{\link[pcaMethods:pca]{pcaMethods::pca()}}.
Impute missing values using singular value decomposition (SVD) imputation.
SVD is a matrix factorization technique that factors a matrix into three matrices:
U, Σ, and V. SVD is used to find the best lower rank approximation of the original matrix.
}
