% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{impute_svd}
\alias{impute_svd}
\title{SVD Imputation}
\usage{
impute_svd(exp, by = NULL, ...)
}
\arguments{
\item{exp}{An expression matrix.}

\item{by}{A grouping variable to consider when imputing missing values.
This variable should be a column in the sample information table.}

\item{...}{Additional arguments to pass to \code{pcaMethods::pca()}.}
}
\value{
An expression matrix with missing values imputed.
}
\description{
A wrapper around the \code{\link[pcaMethods:pca]{pcaMethods::pca()}}.
Impute missing values using singular value decomposition (SVD) imputation.
SVD is a matrix factorization technique that factors a matrix into three matrices:
U, Σ, and V. SVD is used to find the best lower rank approximation of the original matrix.
}
