% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{impute_zero}
\alias{impute_zero}
\title{Zero Imputation}
\usage{
impute_zero(x)
}
\arguments{
\item{x}{A \code{\link[glyexp:experiment]{glyexp::experiment()}} object. \code{\link[glyexp:GlycomicSE]{glyexp::GlycomicSE()}} and
\code{\link[glyexp:GlycoproteomicSE]{glyexp::GlycoproteomicSE()}} objects are also supported.}
}
\value{
A \code{\link[glyexp:experiment]{glyexp::experiment()}} object with missing values imputed. Glyco
SE inputs return the same subclass.
}
\description{
Impute missing values in an expression matrix by replacing them with zeros.
}
