% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{impute_zero}
\alias{impute_zero}
\title{Zero Imputation}
\usage{
impute_zero(x)
}
\arguments{
\item{x}{A \code{\link[glyexp:GlycomicSE]{glyexp::GlycomicSE()}}, \code{\link[glyexp:GlycoproteomicSE]{glyexp::GlycoproteomicSE()}}, or
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment::SummarizedExperiment()}} object.}
}
\value{
A container of the same class as \code{x}, with missing values imputed.
}
\description{
Impute missing values in an expression matrix by replacing them with zeros.
}
