% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{loesscyc_normalize}
\alias{loesscyc_normalize}
\title{LoessCyc Normalization}
\usage{
loesscyc_normalize(exp, ...)
}
\arguments{
\item{exp}{An experiment object.}

\item{...}{Additional arguments to pass to \code{\link[limma:normalizeCyclicLoess]{limma::normalizeCyclicLoess()}}.}
}
\value{
An experiment object with the expression matrix normalized.
}
\description{
This function is a wrapper around \code{\link[limma:normalizeCyclicLoess]{limma::normalizeCyclicLoess()}} with
\code{method = "pairs"}.
Each pair of columns is normalized mutually to each other.
See \href{https://doi.org/10.1093/bib/bbw095}{this paper} for more information.
Also see \code{\link[limma:normalizeCyclicLoess]{limma::normalizeCyclicLoess()}}.
}
