% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize_median}
\alias{normalize_median}
\title{Median Normalization}
\usage{
normalize_median(x)
}
\arguments{
\item{x}{Either a \code{glyexp_experiment} object or a matrix.
If a matrix, rows should be variables and columns should be samples.}
}
\value{
Returns the same type as the input. If \code{x} is a \code{glyexp_experiment},
returns a \code{glyexp_experiment} with normalized expression matrix.
If \code{x} is a matrix, returns a normalized matrix.
}
\description{
Normalize the expression matrix by dividing each column (sample) by the median
of that column (NA ignored), so that the median of each column is 1.
This is the most common normalization method for proteomics data.
It effectively and robustly removes the bias introduced by total protein abundance,
and removes batch effects in part.
}
