% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize_median_abs}
\alias{normalize_median_abs}
\title{Absolute Median Normalization}
\usage{
normalize_median_abs(x)
}
\arguments{
\item{x}{A \code{\link[glyexp:experiment]{glyexp::experiment()}} or
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment::SummarizedExperiment()}} object.}
}
\value{
A \code{\link[glyexp:experiment]{glyexp::experiment()}} object with a normalized expression matrix.
SummarizedExperiment inputs return the same class.
}
\description{
Normalize the expression matrix by dividing each column (sample) by the absolute
median of that column (NA ignored), so that the absolute median of each column is 1.
}
