% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize_median_abs}
\alias{normalize_median_abs}
\title{Absolute Median Normalization}
\usage{
normalize_median_abs(x)
}
\arguments{
\item{x}{A \code{\link[glyexp:experiment]{glyexp::experiment()}} object. \code{\link[glyexp:GlycomicSE]{glyexp::GlycomicSE()}} and
\code{\link[glyexp:GlycoproteomicSE]{glyexp::GlycoproteomicSE()}} objects are also supported.}
}
\value{
A \code{\link[glyexp:experiment]{glyexp::experiment()}} object with a normalized expression matrix.
Glyco SE inputs return the same subclass.
}
\description{
Normalize the expression matrix by dividing each column (sample) by the absolute
median of that column (NA ignored), so that the absolute median of each column is 1.
}
