% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize_median_abs}
\alias{normalize_median_abs}
\title{Absolute Median Normalization}
\usage{
normalize_median_abs(exp)
}
\arguments{
\item{exp}{An experiment object.}
}
\value{
An experiment object with the expression matrix normalized.
}
\description{
Normalize the expression matrix by dividing each column (sample) by the absolute
median of that column (NA ignored), so that the absolute median of each column is 1.
}
