% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize_median_abs}
\alias{normalize_median_abs}
\alias{normalize_median_abs.glyexp_experiment}
\alias{normalize_median_abs.matrix}
\alias{normalize_median_abs.default}
\title{Absolute Median Normalization}
\usage{
normalize_median_abs(x)

\method{normalize_median_abs}{glyexp_experiment}(x)

\method{normalize_median_abs}{matrix}(x)

\method{normalize_median_abs}{default}(x)
}
\arguments{
\item{x}{Either a \code{glyexp_experiment} object or a matrix.
If a matrix, rows should be variables and columns should be samples.}
}
\value{
Returns the same type as the input. If \code{x} is a \code{glyexp_experiment},
returns a \code{glyexp_experiment} with normalized expression matrix.
If \code{x} is a matrix, returns a normalized matrix.
}
\description{
Normalize the expression matrix by dividing each column (sample) by the absolute
median of that column (NA ignored), so that the absolute median of each column is 1.
}
