% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize_quantile}
\alias{normalize_quantile}
\title{Quantile Normalization}
\usage{
normalize_quantile(exp, by = NULL, ...)
}
\arguments{
\item{exp}{An experiment object.}

\item{by}{A column name in \code{sample_info} to stratify by. Optional.
If provided, the normalization will be performed within each group.}

\item{...}{Additional arguments to pass to \code{\link[limma:normalizequantiles]{limma::normalizeQuantiles()}}.}
}
\value{
An experiment object with the expression matrix normalized.
}
\description{
This function is a wrapper around \code{\link[limma:normalizequantiles]{limma::normalizeQuantiles()}}.
It normalizes the expression matrix by quantile normalization.
This method is used to remove technical variation between samples.
Proteomics data rarely uses this method, but it is common in microarray data.
See \href{https://en.wikipedia.org/wiki/Quantile_normalization}{wikipedia}
for more information.
}
