% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize_rlrma}
\alias{normalize_rlrma}
\title{Robust Linear Regression with Median Adjustment Normalization}
\usage{
normalize_rlrma(exp, by = NULL)
}
\arguments{
\item{exp}{An experiment object.}

\item{by}{A column name in \code{sample_info} to stratify by. Optional.
If provided, the normalization will be performed within each group.}
}
\value{
An experiment object with the expression matrix normalized.
}
\description{
This method is based on robust linear regression with median adjustment.
First, the median of each variable's abundance across all measured samples
is subtracted from the sample's abundance values. Then, it's like the
\code{\link[=normalize_rlr]{normalize_rlr()}} method.
}
