% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize_rlrma}
\alias{normalize_rlrma}
\title{Robust Linear Regression with Median Adjustment Normalization}
\usage{
normalize_rlrma(x, by = NULL)
}
\arguments{
\item{x}{Either a \code{glyexp_experiment} object or a matrix.
If a matrix, rows should be variables and columns should be samples.}

\item{by}{A column name in \code{sample_info} to stratify by. Optional.
If provided, the normalization will be performed within each group.
Only available for \code{glyexp_experiment} objects.}
}
\value{
Returns the same type as the input. If \code{x} is a \code{glyexp_experiment},
returns a \code{glyexp_experiment} with normalized expression matrix.
If \code{x} is a matrix, returns a normalized matrix.
}
\description{
This method is based on robust linear regression with median adjustment.
First, the median of each variable's abundance across all measured samples
is subtracted from the sample's abundance values. Then, it's like the
\code{\link[=normalize_rlr]{normalize_rlr()}} method.
}
