% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize_total_area}
\alias{normalize_total_area}
\alias{normalize_total_area.glyexp_experiment}
\alias{normalize_total_area.matrix}
\alias{normalize_total_area.default}
\title{Total Area Normalization}
\usage{
normalize_total_area(x)

\method{normalize_total_area}{glyexp_experiment}(x)

\method{normalize_total_area}{matrix}(x)

\method{normalize_total_area}{default}(x)
}
\arguments{
\item{x}{Either a \code{glyexp_experiment} object or a matrix.
If a matrix, rows should be variables and columns should be samples.}
}
\value{
Returns the same type as the input. If \code{x} is a \code{glyexp_experiment},
returns a \code{glyexp_experiment} with normalized expression matrix.
If \code{x} is a matrix, returns a normalized matrix.
}
\description{
Normalize the expression matrix by dividing each column (sample) by the sum of
that column, so that the sum of each column is 1.
This is the most common normalization method for glycomics data.
It removes the bias introduced by total glycan abundance.
However, it results in compositional data, which may result in unrealistic
downstream analysis results.
}
