% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize_vsn}
\alias{normalize_vsn}
\title{Variance Stabilizing Normalization}
\usage{
normalize_vsn(exp, by = NULL, ...)
}
\arguments{
\item{exp}{An experiment object.}

\item{by}{A column name in \code{sample_info} to stratify by. Optional.
If provided, the normalization will be performed within each group.}

\item{...}{Additional arguments to pass to \code{\link[limma:normalizeVSN]{limma::normalizeVSN()}}.}
}
\value{
An experiment object with the expression matrix normalized.
}
\description{
This function is a wrapper around \code{\link[limma:normalizeVSN]{limma::normalizeVSN()}}.
Evidence proved that this method performs well in reducing noise and
boosting differential expression detection.
Log-transformation is not needed for downstream statistical analysis,
for this normalization method performs a log-like transformation internally.
Due to the same reason, fold change estimates will be severely distorted.
Please use this method with caution.
See \href{https://doi.org/10.1093/bib/bbw095}{this paper} for more information.
}
\details{
At least 42 variables should be provided for this method.
This follows the convention of the \code{vsn} package.
}
