% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{plot_tic_bar}
\alias{plot_tic_bar}
\title{Plot Total Intensity by Sample}
\usage{
plot_tic_bar(exp)
}
\arguments{
\item{exp}{A \code{\link[glyexp:experiment]{glyexp::experiment()}} object.}
}
\value{
A ggplot object of total intensity by sample.
}
\description{
Draw a bar plot showing total intensity (TIC) for each sample. Samples are
ordered from high to low TIC from left to right.
}
\examples{
plot_tic_bar(glyexp::toy_experiment)

}
