% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.R
\name{remove_low_cv}
\alias{remove_low_cv}
\title{Remove Variables with Low Coefficient of Variation}
\usage{
remove_low_cv(x, cv_cutoff = 0, by = NULL, strict = FALSE)
}
\arguments{
\item{x}{A \code{\link[glyexp:experiment]{glyexp::experiment()}} object. \code{\link[glyexp:GlycomicSE]{glyexp::GlycomicSE()}} and
\code{\link[glyexp:GlycoproteomicSE]{glyexp::GlycoproteomicSE()}} objects are also supported.}

\item{cv_cutoff}{The cutoff for coefficient of variation. Defaults to 0.}

\item{by}{A factor specifying the groupings. Defaults to NULL.}

\item{strict}{If \code{FALSE}, remove a variable only if it passes the coefficient of variation threshold in all groups.
If \code{TRUE}, remove a variable if it passes the coefficient of variation threshold in any group.}
}
\value{
A \code{\link[glyexp:experiment]{glyexp::experiment()}} object with filtered variables. Glyco SE
inputs return the same subclass.
}
\description{
Filters variables whose coefficient of variation falls below a threshold.
Default behavior is to remove variables with zero coefficient of variation.
}
\seealso{
\code{\link[=remove_low_var]{remove_low_var()}}
}
