% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.R
\name{remove_low_cv}
\alias{remove_low_cv}
\alias{remove_low_cv.glyexp_experiment}
\alias{remove_low_cv.matrix}
\alias{remove_low_cv.default}
\title{Remove Variables with Low Coefficient of Variation}
\usage{
remove_low_cv(x, cv_cutoff = 0, by = NULL, strict = FALSE)

\method{remove_low_cv}{glyexp_experiment}(x, cv_cutoff = 0, by = NULL, strict = FALSE)

\method{remove_low_cv}{matrix}(x, cv_cutoff = 0, by = NULL, strict = FALSE)

\method{remove_low_cv}{default}(x, cv_cutoff = 0, by = NULL, strict = FALSE)
}
\arguments{
\item{x}{Either a \code{glyexp_experiment} object or a matrix.}

\item{cv_cutoff}{The cutoff for coefficient of variation. Defaults to 0.}

\item{by}{A factor specifying the groupings. Defaults to NULL.}

\item{strict}{If \code{FALSE}, remove a variable only if it passes the coefficient of variation threshold in all groups.
If \code{TRUE}, remove a variable if it passes the coefficient of variation threshold in any group.}
}
\value{
For \code{glyexp_experiment} input, returns a modified \code{glyexp_experiment} object.
For matrix input, returns a filtered matrix.
}
\description{
Filters variables whose coefficient of variation falls below a threshold.
Default behavior is to remove variables with zero coefficient of variation.
}
\seealso{
\code{\link[=remove_low_var]{remove_low_var()}}
}
