% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.R
\name{remove_low_expr}
\alias{remove_low_expr}
\alias{remove_low_expr.glyexp_experiment}
\alias{remove_low_expr.matrix}
\alias{remove_low_expr.default}
\title{Remove Variables with Low Expression}
\usage{
remove_low_expr(x, percentile = 0.05, by = NULL, strict = FALSE)

\method{remove_low_expr}{glyexp_experiment}(x, percentile = 0.05, by = NULL, strict = FALSE)

\method{remove_low_expr}{matrix}(x, percentile = 0.05, by = NULL, strict = FALSE)

\method{remove_low_expr}{default}(x, percentile = 0.05, by = NULL, strict = FALSE)
}
\arguments{
\item{x}{Either a \code{glyexp_experiment} object or a matrix.}

\item{percentile}{The percentile for median expression values.
Defaults to 0.05, i.e., the 5\% lowest median expression values will be removed.}

\item{by}{Either a column name in \code{sample_info} (string) or a vector specifying group assignments for each sample.}

\item{strict}{If \code{FALSE}, remove a variable only if it passes the abundance thresholds in all groups.
If \code{TRUE}, remove a variable if it passes the abundance thresholds in any group. Defaults to FALSE.}
}
\value{
For \code{glyexp_experiment} input, returns a modified \code{glyexp_experiment} object.
For matrix input, returns a filtered matrix.
}
\description{
Filters variables based on median expression values.
Variables with median expression values below certain percentile will be removed.
}
